feat: update env & remove login test(login will update)
