+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "BusinessMonitor",
-  "version": "0.1.0+114",
+  "version": "0.1.0+115",
   "private": true,
   "type": "module",
   "scripts": {
