+++ src/constants/dewt.ts
@@ -1,0 +1,7 @@
+export const RPID = process.env.RPID;
+export const ORIGIN = process.env.ORIGIN;
+export const DEWT_ISS = process.env.DEWT_ISS;
+export const DEWT_AUD = process.env.DEWT_AUD;
+export const DEWT_MAX_AGE_SEC = process.env.DEWT_MAX_AGE_SEC;
+export const DEWT_KID = process.env.DEWT_KID;
+export const DEWT_JWK = process.env.DEWT_JWK;
